Move BlockaidIndicator into settings Change i18n name
