+++ app/components/Views/Settings/SecuritySettings/BlockaidIndicator/index.ts
