+++ app/core/Analytics/MetaMetrics.events.ts
@@ -800,9 +800,7 @@ const events = {
   ),
 
   // Experimental Settings
-  SETTINGS_EXPERIMENTAL_SECURITY_ALERTS_ENABLED: generateOpt(
-    EVENT_NAME.SETTINGS_UPDATED,
-  ),
+  SETTINGS_SECURITY_ALERTS_ENABLED: generateOpt(EVENT_NAME.SETTINGS_UPDATED),
 
   // Ledger
   CONNECT_LEDGER: generateOpt(EVENT_NAME.CONNECT_LEDGER),
