+++ locales/languages/hi.json
@@ -888,6 +888,10 @@
     "network_provider": "नेटवर्क प्रोवाइडर",
     "token_nft_ens_subheading": "टोकन, NFT और ENS ऑटो-डिटेक्शन",
     "security_check_subheading": "सुरक्षा जाँचें",
+    "security_alerts": "सुरक्षा चेतावनियाँ",
+    "security_alerts_desc": "यह सुविधा स्थानीय रूप से आपके ट्रांसेक्शन और हस्ताक्षर अनुरोधों की समीक्षा करके आपको बुरी नीयत वाली गतिविधि के प्रति एलर्ट करती है। किसी भी अनुरोध को मंजूरी देने से पहले हमेशा पूरी जांच-पड़ताल ज़रूर करें। इस बात की कोई गारंटी नहीं है कि यह सुविधा सभी बुरी नीयत वाली गतिविधि का पता लगा लेगी। इस सुविधा को सक्षम करके आप प्रदाता की उपयोग की शर्तों से सहमत होते हैं।",    
+    "blockaid": "Blockaid",
+    "blockaid_desc": "गोपनीयता को सुरक्षित रखना - कोई भी डेटा थर्ड पार्टी के साथ शेयर नहीं किया जाता है। Arbitrum, Avalanche, BNB chain, Ethereum Mainnet, Optimism और Polygon पर उपलब्ध है।",    
     "symbol_required": "Symbol is required."
   },
   "sdk": {
@@ -1841,11 +1845,7 @@
     "wallet_connect_dapps_desc": "सक्रिय वॉलेटकनेक्ट सत्रों की सूची देखें",
     "wallet_connect_dapps_cta": "सेशन देखें",
     "network_not_supported": "वर्तमान नेटवर्क समर्थन नहीं करता",
-    "security_alerts": "सुरक्षा चेतावनियाँ",
-    "security_alerts_desc": "यह सुविधा स्थानीय रूप से आपके ट्रांसेक्शन और हस्ताक्षर अनुरोधों की समीक्षा करके आपको बुरी नीयत वाली गतिविधि के प्रति एलर्ट करती है। किसी भी अनुरोध को मंजूरी देने से पहले हमेशा पूरी जांच-पड़ताल ज़रूर करें। इस बात की कोई गारंटी नहीं है कि यह सुविधा सभी बुरी नीयत वाली गतिविधि का पता लगा लेगी। इस सुविधा को सक्षम करके आप प्रदाता की उपयोग की शर्तों से सहमत होते हैं।",
     "select_provider": "अपना पसंदीदा प्रदाता चुनें",
-    "blockaid": "Blockaid",
-    "blockaid_desc": "गोपनीयता को सुरक्षित रखना - कोई भी डेटा थर्ड पार्टी के साथ शेयर नहीं किया जाता है। Arbitrum, Avalanche, BNB chain, Ethereum Mainnet, Optimism और Polygon पर उपलब्ध है।",
     "switch_network": "कृपया Mainnet या sepolia पर बदलें"
   },
   "walletconnect_sessions": {
