+++ locales/languages/ru.json
@@ -888,6 +888,10 @@
     "network_provider": "Поставщик услуг сети",
     "token_nft_ens_subheading": "Автоопределение токенов, NFT и ENS",
     "security_check_subheading": "Проверки безопасности",
+    "security_alerts": "Оповещения безопасности",
+    "security_alerts_desc": "Эта функция предупреждает вас о вредоносной активности, проверяя запросы транзакций и подписей локально. Всегда проводите комплексную проверку перед утверждением каких-либо запросов. Нет никакой гарантии, что эта функция обнаружит всю вредоносную активность. Включая эту функцию, вы соглашаетесь с условиями использования поставщика.",    
+    "blockaid": "Blockaid",
+    "blockaid_desc": "Сохранение конфиденциальности – никакие данные не передаются третьим сторонам. Доступно в Arbitrum, Avalanche, BNB Chain, Мейн-нете Ethereum, Optimism и Polygon.",    
     "symbol_required": "Symbol is required."
   },
   "sdk": {
@@ -1841,11 +1845,7 @@
     "wallet_connect_dapps_desc": "Просмотр списка активных сеансов WalletConnect",
     "wallet_connect_dapps_cta": "Просмотр сеансов",
     "network_not_supported": "Текущая сеть не поддерживается",
-    "security_alerts": "Оповещения безопасности",
-    "security_alerts_desc": "Эта функция предупреждает вас о вредоносной активности, проверяя запросы транзакций и подписей локально. Всегда проводите комплексную проверку перед утверждением каких-либо запросов. Нет никакой гарантии, что эта функция обнаружит всю вредоносную активность. Включая эту функцию, вы соглашаетесь с условиями использования поставщика.",
     "select_provider": "Выберите предпочитаемого поставщика",
-    "blockaid": "Blockaid",
-    "blockaid_desc": "Сохранение конфиденциальности – никакие данные не передаются третьим сторонам. Доступно в Arbitrum, Avalanche, BNB Chain, Мейн-нете Ethereum, Optimism и Polygon.",
     "switch_network": "Переключитесь на мейн-нет или Sepolia"
   },
   "walletconnect_sessions": {
